# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Generic SNAP/ASPI-compliant payment gateway API: B2B access\ntoken issuance, admin client onboarding, signature-generation\nutilities, and Virtual Account (transfer-VA) transaction and\nmerchant-dashboard endpoints.",
        "title": "SNAP Payment Integration Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/openapi/v1.0/transfer-va/inquiry": {
            "post": {
                "security": [
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Vendor-initiated inquiry for Virtual Account bill/customer details prior to payment. Read-only.",
                "tags": [
                    "Virtual Account"
                ],
                "summary": "VA bill inquiry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Symmetric signature; compute via POST /api/v1/utilities/signature-service",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique external ID for this request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "VA inquiry request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Field Format / Invalid Mandatory Field / missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: request already in progress for this Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.VAInquiryResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/v1.0/transfer-va/payment": {
            "post": {
                "security": [
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Vendor-initiated notification that a payment against a Virtual Account has been received. State-changing: records the payment.",
                "tags": [
                    "Virtual Account"
                ],
                "summary": "VA payment notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Symmetric signature; compute via POST /api/v1/utilities/signature-service",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique external ID for this request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "VA payment notification",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Field Format / Invalid Mandatory Field / missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (mapped from downstream error, or in-flight request with same Idempotency-Key)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.VAPaymentResponse"
                        }
                    }
                }
            }
        },
        "/openapi/v1.0/transfer-va/status": {
            "post": {
                "security": [
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Vendor-initiated inquiry of the current payment status of a Virtual Account transaction. Read-only.",
                "tags": [
                    "Virtual Account"
                ],
                "summary": "VA payment status inquiry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Symmetric signature; compute via POST /api/v1/utilities/signature-service",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique external ID for this request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "VA status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Field Format / Invalid Mandatory Field / missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: request already in progress for this Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.VAStatusResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Amount": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "domain.BilingualText": {
            "type": "object",
            "properties": {
                "english": {
                    "type": "string"
                },
                "indonesia": {
                    "type": "string"
                }
            }
        },
...
        "domain.VAAccountData": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "billDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BillDetail"
                    }
                },
                "customerNo": {
                    "type": "string"
                },
                "feeAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "freeTexts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BilingualText"
                    }
                },
                "inquiryReason": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "inquiryRequestId": {
                    "type": "string"
                },
                "inquiryStatus": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "subCompany": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "virtualAccountName": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                },
                "virtualAccountTrxType": {
                    "type": "string"
                }
            }
        },
        "domain.VAInquiryRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "amount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "channelCode": {
                    "type": "integer"
                },
                "customerNo": {
                    "type": "string"
                },
                "inquiryRequestId": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "txnDateInit": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.VAInquiryResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "virtualAccountData": {
                    "$ref": "#/definitions/domain.VAAccountData"
                }
            }
        },
...
        "domain.VAPaymentBillDetail": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "billAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "billCode": {
                    "type": "string"
                },
                "billDescription": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "billName": {
                    "type": "string"
                },
                "billNo": {
                    "type": "string"
                },
                "billReferenceNo": {
                    "type": "string"
                },
                "billShortName": {
                    "type": "string"
                },
                "billSubCompany": {
                    "type": "string"
                },
                "reason": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "domain.VAPaymentRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "billDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.VAPaymentBillDetail"
                    }
                },
                "customerNo": {
                    "type": "string"
                },
                "flagAdvise": {
                    "type": "string"
                },
                "freeTexts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BilingualText"
                    }
                },
                "inquiryRequestId": {
                    "type": "string"
                },
                "paidAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "paidBills": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "paymentRequestId": {
                    "type": "string"
                },
                "paymentType": {
                    "type": "string"
                },
                "referenceNo": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "trxDateTime": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.VAPaymentResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "virtualAccountData": {
                    "$ref": "#/definitions/domain.VAPaymentStatus"
                }
            }
        },
        "domain.VAPaymentStatus": {
            "type": "object",
            "properties": {
                "customerNo": {
                    "type": "string"
                },
                "paidAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "paymentFlagReason": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "paymentFlagStatus": {
                    "type": "string"
                },
                "paymentRequestId": {
                    "type": "string"
                },
                "referenceNo": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "trxDateTime": {
                    "type": "string"
                },
                "trxId": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
...
        "domain.VAStatusData": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "billDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.VAStatusBillDetail"
                    }
                },
                "customerNo": {
                    "type": "string"
                },
                "freeTexts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BilingualText"
                    }
                },
                "inquiryRequestId": {
                    "type": "string"
                },
                "paidAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "paidBills": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "paymentFlagReason": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "paymentFlagStatus": {
                    "type": "string"
                },
                "paymentRequestId": {
                    "type": "string"
                },
                "paymentType": {
                    "type": "string"
                },
                "referenceNo": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "transactionDate": {
                    "type": "string"
                },
                "trxDateTime": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.VAStatusRequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customerNo": {
                    "type": "string"
                },
                "inquiryRequestId": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.VAStatusResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "virtualAccountData": {
                    "$ref": "#/definitions/domain.VAStatusData"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "Static admin API key configured via the ADMIN_API_KEY environment variable. Required for all /admin/* endpoints.",
            "type": "apiKey",
            "name": "X-Admin-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer access token issued by POST /openapi/v1.0/access-token/b2b, e.g. \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SnapClientKey": {
            "description": "Client identifier issued by the admin during onboarding. Required on the B2B access-token endpoint.",
            "type": "apiKey",
            "name": "X-CLIENT-KEY",
            "in": "header"
        },
        "SnapSignature": {
            "description": "Request signature. Compute via POST /api/v1/utilities/signature-auth (asymmetric, access-token endpoint) or POST /api/v1/utilities/signature-service (symmetric, transaction endpoints).",
            "type": "apiKey",
            "name": "X-SIGNATURE",
            "in": "header"
        },
        "SnapTimestamp": {
            "description": "Request timestamp in ISO 8601 format (e.g. 2026-07-24T10:00:00+07:00).",
            "type": "apiKey",
            "name": "X-TIMESTAMP",
            "in": "header"
        }
    }
}